refactor: add utils.js
