+++ src/main.js
@@ -1,18 +1,12 @@
 const { app, BrowserWindow } = require('electron')
+const { initWindow } = require('./utils')
 require('./getData')
 require('./excel')
 
 const isDev = !app.isPackaged
-let win = null
 
 function createWindow () {
-  win = new BrowserWindow({
-    width: 888,
-    height: 550,
-    webPreferences: {
-      nodeIntegration: true
-    }
-  })
+  const win = initWindow()
   win.setMenuBarVisibility(false)
   isDev ? win.loadURL(`http://localhost:3000`) : win.loadFile('./dist/index.html')
 }
@@ -30,7 +24,3 @@ app.on('activate', () => {
     createWindow()
   }
 })
-
-const getWin = () => win
-
-exports.getWin = getWin
